feat: multi target publishing for car app target
