+++ buildSrc/.gitignore
@@ -1,0 +1,12 @@
+*.iml
+.gradle
+/local.properties
+/.idea
+/.kotlin
+.DS_Store
+/build
+/captures
+.externalNativeBuild
+.cxx
+local.properties
+api_keys.xml
