+++ buildSrc/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    `kotlin-dsl`
+}
+
+repositories {
+    google()
+    mavenCentral()
+    gradlePluginPortal()
+}
