+++ buildSrc/src/main/resources/META-INF/gradle-plugins/CommonPomConventionPlugin.properties
@@ -1,0 +1,1 @@
+implementation-class=CommonPomConventionPlugin
